+++ src/main/kotlin/dev/slint/ideaplugin/ide/lsp/LspLanguageClient.kt
@@ -1,6 +1,7 @@
 package dev.slint.ideaplugin.ide.lsp
 
 import com.intellij.openapi.command.WriteCommandAction
+import com.intellij.openapi.components.service
 import com.intellij.openapi.editor.Editor
 import com.intellij.openapi.editor.LogicalPosition
 import com.intellij.openapi.fileEditor.FileEditorManager
@@ -8,6 +9,8 @@ import com.intellij.openapi.fileEditor.OpenFileDescriptor
 import com.intellij.openapi.project.Project
 import com.intellij.platform.lsp.api.Lsp4jClient
 import com.intellij.platform.lsp.api.LspServerNotificationsHandler
+import dev.slint.ideaplugin.ide.services.FileEditorService
+import dev.slint.ideaplugin.ide.services.SlintServerService
 import org.eclipse.lsp4j.*
 import org.eclipse.lsp4j.jsonrpc.services.JsonNotification
 import java.util.concurrent.CompletableFuture
@@ -24,7 +27,7 @@ class LspServerNotificationsMiddleware(
     project: Project
 ): LspServerNotificationsHandler {
 
-    private val fileEditorService = FileEditorService(project)
+    private val fileEditorService = project.service<FileEditorService>()
 
     override fun applyEdit(params: ApplyWorkspaceEditParams): CompletableFuture<ApplyWorkspaceEditResponse> {
         fileEditorService.applyEdit(params)
@@ -110,72 +113,5 @@ class LspServerNotificationsMiddleware(
 
     override fun workspaceFolders(): CompletableFuture<List<WorkspaceFolder>> {
         return serverNotificationsHandler.workspaceFolders()
-    }
-}
-
-class FileEditorService(private val project: Project) {
-    private val fileEditorManager = FileEditorManager.getInstance(project)
-
-    fun applyEdit(params: ApplyWorkspaceEditParams) {
-        params.edit.documentChanges.forEach {
-            if (it.isLeft) {
-                applyDocumentChanges(it.left)
-            }
-        }
-    }
-
-    fun showDocument(params: ShowDocumentParams) {
-        fileEditorManager.openFiles
-            .find {
-                it.url == params.uri
-            }
-            ?.let {
-                val startPosition = params.selection.start.run {
-                    LogicalPosition(line, character)
-                }
-                val endPosition = params.selection.end.run {
-                    LogicalPosition(line, character)
-                }
-
-                WriteCommandAction.runWriteCommandAction(project) {
-                    fileEditorManager
-                        .openTextEditor(OpenFileDescriptor(project, it), true)
-                        ?.selectionModel
-                        ?.setBlockSelection(startPosition, endPosition)
-                }
-            }
-    }
-
-    private fun applyDocumentChanges(documentChanges: TextDocumentEdit) {
-        val fileUrl = documentChanges.textDocument.uri
-        fileEditorManager.openFiles
-            .find {
-                it.url == fileUrl
-            }
-            ?.let { virtualFile ->
-                WriteCommandAction.runWriteCommandAction(project) {
-                    fileEditorManager
-                        .openTextEditor(OpenFileDescriptor(project, virtualFile), true)
-                        ?.let { editor: Editor ->
-                            documentChanges.edits.forEach {
-                                applyTextChanges(it, editor)
-                            }
-                        }
-                }
-            }
-    }
-
-    private fun applyTextChanges(textEdit: TextEdit, editor: Editor) {
-        val startPosition = textEdit.range.start.run {
-            LogicalPosition(line, character)
-        }
-        val endPosition = textEdit.range.end.run {
-            LogicalPosition(line, character)
-        }
-
-        val startOffset = editor.logicalPositionToOffset(startPosition)
-        val endOffset= editor.logicalPositionToOffset(endPosition)
-
-        editor.document.replaceString(startOffset, endOffset, textEdit.newText)
     }
 }