+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Widget for indicating the LSP-server current status
 - Pick elements with your mouse by enabling the "Design Mode"
 - Drag and Drop on Live-Preview
+- Document formatting support
 
 ### Fixed